requirements.txt
numpy==1.19.2
Pillow==8.1.0
streamlit==1.5.0
tensorflow-cpu==2.7.0
